Add class for recording changes to board (Moves)
An inner class for SudokuBoard, Move, was added to track changes
made to a board. Moves will be useful when describing the order of
changes needed to solve a board. Moves are added whenever a change
is made to the value or pencils on the board. Moves can be undone
in a stack-like way in which recently made moves are removed.
Appropriate test methods were added for the new methods in
SudokuBoard.