Add two algorithms for solving puzzles
The first two algorithms/techniques for solving puzzles were
added, as well as a method for applying these methods until the board
is solved. singlePosition finds the cell in a row/column/region that
is the only cell in the group that can have a value. singleCandidate
finds cells that can only have one value (all other possibilities are
eliminated). More algorithms will be added to make the solve method
more robust.