Optimize puzzle solving algorithms
The brute force solving algorithm has been optimized by first sending the
puzzle through the deterministic solving methods. These methods effectively
reduce the size of the problem that the brute force algorithm has to solve.
When a puzzle is passed through the deterministic methods, it is likely that
some of the empty squares will be filled with values, reducing the problem
size for the brute force solver.

The deterministic solver was also optimized. The loop that handles the
iterations of the solving methods will now terminate if the board has not
changed by the end of the current iteration. This results in a massively
faster deterministic solve as well as a faster optimized brute force solve.

Speed comparisons between the optimized and unoptimized version are highly
dependent on the input puzzles. Tests on samples of 100 random puzzles show
between 2.25 to 27.5 times faster average solve speed.