Make all SudokuSolver methods static
The SudokuSolver class does not contain any fields, and as a result
there is no way of differentiating one SudokuSolver instance from
another because it is only used to access its methods. Thus all of the
methods were made static so that an instance of SudokuSolver does not
have to be created in order to access its methods. Other code affected
by this change was updated accordingly.